+++ manager-web/src/main/java/com/shengde/e3mall/controller/ItemCatController.java
@@ -21,7 +21,7 @@ public class ItemCatController {
 
     @RequestMapping("/item/cat/list")
     @ResponseBody
-    public List<EasyUITreeNode> getItemCatList(@RequestParam(name = "id",defaultValue = "0") Long parentId){
+    public List<EasyUITreeNode> getItemCatList(@RequestParam(name = "id",defaultValue = "0") int parentId){
 
         //调用服务查询节点列表
         List<EasyUITreeNode> list = itemCatService.getItemCatlist(parentId);
